refactor: изменение структуры приложения (#1)
Изменение структуры форм данных
Вынос цветов в переменные
