+++ frontend/src/components/UI/Input/Input.module.scss
@@ -1,6 +1,6 @@
 input {
 	border-radius: 8px;
-	background: #fff;
+	background: var(--input-background-color);
 	width: 462px;
 	height: 27px;
 	padding: 13px 14px;
@@ -9,20 +9,25 @@ input {
 }
 
 .input_valid {
-	border: 1px solid #e3e3e3;
+	border: 1px solid var(--input-valid-border-color);
 }
 
 .input_invalid {
-	border: 1px solid #e83434;
+	border: 1px solid var(--input-invalid-border-color);
 }
 
 input::placeholder {
-	color: #c1c1c1;
+	color: var(--input-placeholder-color);
 }
 
-input:focus,
+input:focus {
+	border: 1px solid var(--input-focus-border-color);
+	box-shadow: 0px 0px 0px 2px rgba(var(--input-focus-box-shadow-color));
+	outline: none;
+}
+
 input:hover {
-	border: 1px solid #bbb;
-	box-shadow: 0px 0px 0px 2px rgba(0, 0, 0, 0.05);
+	border: 1px solid var(--input-hover-border-color);
+	box-shadow: 0px 0px 0px 2px rgba(var(--input-hover-box-shadow-color));
 	outline: none;
 }
