+++ frontend/src/pages/CheckEmail/CheckEmail.module.scss
@@ -9,7 +9,7 @@
 		margin-top: 184px;
 
 		.check_email__title {
-			color: #000;
+			color: var(--primary-text-color);
 			font-size: 32px;
 			font-weight: 700;
 		}
@@ -17,7 +17,7 @@
 		.check_email__text {
 			margin: 60px auto;
 			width: 624px;
-			color: #000;
+			color: var(--primary-text-color);
 			text-align: center;
 			font-size: 24px;
 			font-weight: 400;
