+++ frontend/src/pages/CheckEmail/index.ts
