+++ frontend/src/pages/PasswordReset/PasswordReset.module.scss
@@ -1,13 +1,12 @@
-.login_container {
+.password_reset_container {
 	width: 490px;
-	background: #fff;
 	position: flex;
 	text-align: center;
 	margin-right: auto;
 	margin-left: auto;
 
-	.login_title {
-		color: #000;
+	.password_reset_title {
+		color: var(--primary-text-color);
 		font-size: 32px;
 		font-weight: 700;
 		margin-top: 0px;
@@ -15,22 +14,22 @@
 		margin-top: 120px;
 	}
 
-	.login_input__group {
+	.password_reset_input__group {
 		margin-top: 16px;
 		text-align: left;
 	}
 
-	.login_label {
-		color: #000;
+	.password_reset_label {
+		color: var(--primary-text-color);
 		font-size: 16px;
 		font-weight: 400;
 		display: block;
 		margin-bottom: 8px;
 	}
 
-	.login_error__validation {
+	.password_reset_error__validation {
 		margin-top: 8px;
-		color: #e83434;
+		color: var(--primary-error-text-color);
 		font-size: 16px;
 		font-weight: 300;
 	}
