+++ frontend/src/pages/PasswordResetConfirmed/PasswordResetConfirmed.module.scss
@@ -1,13 +1,12 @@
-.registration_container {
+.password_reset_confirmed_container {
 	width: 490px;
-	background: #fff;
 	position: flex;
 	text-align: center;
 	margin-right: auto;
 	margin-left: auto;
 
-	.registration_title {
-		color: #000;
+	.password_reset_confirmed_title {
+		color: var(--primary-text-color);
 		font-size: 32px;
 		font-weight: 700;
 		margin-top: 0px;
@@ -15,22 +14,22 @@
 		margin-top: 120px;
 	}
 
-	.registration_input__group {
+	.password_reset_confirmed_input__group {
 		margin-top: 16px;
 		text-align: left;
 	}
 
-	.registration_label {
-		color: #000;
+	.password_reset_confirmed_label {
+		color: var(--primary-text-color);
 		font-size: 16px;
 		font-weight: 400;
 		display: block;
 		margin-bottom: 8px;
 	}
 
-	.registration_error__validation {
+	.password_reset_confirmed_error__validation {
 		margin-top: 8px;
-		color: #e83434;
+		color: var(--primary-error-text-color);
 		font-size: 16px;
 		font-weight: 300;
 	}
