+++ frontend/src/pages/PasswordResetConfirmed/index.ts
