+++ frontend/src/pages/Registration/index.ts
